Cleanup and extension of expectations.
EXPECT doesn't need SET variables as the second operand anymore but accepts any
valid memory references.

EXPECT_ANY is new, expects 1 parameter in 'a' and adds a clause OR'ing all
given bits in [a). This is useful for ensuring that two bit-strings differ in
at least one bit position, which is needed for the upcoming collision attack
mode.